TestCluster: borg untilExitsDisconnected from the unad test
